first

+++ src/redux/contacts/operations.js
@@ -1,11 +1,23 @@
 import axios from "axios";
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
-axios.defaults.baseURL = "https://6817a1ad26a599ae7c3af9d9.mockapi.io";
+axios.defaults.baseURL = "https://connections-api.goit.global/";
+
+const setAuthHeader = (token) => {
+  axios.defaults.headers.common.Authorization = `Bearer ${token}`;
+};
 
 export const fetchContacts = createAsyncThunk(
   "contacts/fetchAll",
   async (_, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("No auth token");
+    }
+
+    setAuthHeader(token);
     try {
       const res = await axios.get("/contacts");
       return res.data;
@@ -18,6 +30,14 @@ export const fetchContacts = createAsyncThunk(
 export const addContact = createAsyncThunk(
   "contacts/addContact",
   async (contact, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("No auth token");
+    }
+
+    setAuthHeader(token);
     try {
       const res = await axios.post("/contacts", contact);
       return res.data;
@@ -30,6 +50,14 @@ export const addContact = createAsyncThunk(
 export const deleteContact = createAsyncThunk(
   "contacts/deleteContact",
   async (id, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("No auth token");
+    }
+
+    setAuthHeader(token);
     try {
       await axios.delete(`/contacts/${id}`);
       return id;
